requirements.txt
Package             Version
------------------- -------
asgiref             3.3.4  
Django              3.2.2  
djangorestframework 3.12.4 
isort               5.8.0  
pip                 19.2.3 
pytz                2021.1 
setuptools          41.2.0 
sqlparse            0.4.1  
